Improve perf of map generation + surface becomes integer

+++ Alpha.Core/Provinces/LandProvince.cs
@@ -52,6 +52,6 @@ namespace Alpha.Core.Provinces
         }
 
         public bool IsCoastal { get { return Zones.SelectMany(z => z.Neighbourgs).Any(z => z.Province is SeaProvince); } }
-        public double Surface { get { return Zones.Sum(z => z.Surface); } }
+        public int Surface { get { return (int)Zones.Sum(z => z.Surface); } }
     }
 }
